move out model utils
